requirements.txt
pygame
pysmb
ConfigParser